project module, tasklist module, todo module

+++ src/user/users.service.ts
@@ -4,8 +4,9 @@ import { User } from './entities/user.entity';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
 import * as bcrypt from 'bcrypt';
-import { CompanyInvite } from '../companies/entities/company-invite.entity';
-import { UserCompany } from '../companies/entities/user-company.entity';
+import { CompanyInvite } from '../company/entities/company-invite.entity';
+import { UserCompany } from '../company/entities/user-company.entity';
+import { Company } from '../company/entities/company.entity';
 
 @Injectable()
 export class UsersService {
@@ -15,6 +16,8 @@ export class UsersService {
     private readonly companyInviteRepository: Repository<CompanyInvite>,
     @InjectRepository(UserCompany)
     private readonly userCompanyRepository: Repository<UserCompany>,
+    @InjectRepository(Company)
+    private readonly companyRepository: Repository<Company>,
   ) {}
   async createUser(createUserDto: CreateUserDto): Promise<string> {
     const user: User = new User();
@@ -23,7 +26,7 @@ export class UsersService {
     user.lastname = createUserDto.lastname;
     user.password = await this.hashString(createUserDto.password);
     await this.userRepository.save(user);
-    return 'successfully created user';
+    return `Successfully created user with username "${createUserDto.username}".`;
   }
   async findOneById(id: number) {
     return await this.userRepository.findOne({
@@ -57,15 +60,25 @@ export class UsersService {
     await this.userRepository.delete(user?.id);
     return 'successfully deleted user';
   }
-  findInvitationsToUser(userId: number): Promise<CompanyInvite[]> {
-    return this.companyInviteRepository.find({ where: { sendTo: userId } });
+  async findInvitationsToUser(userId: number): Promise<CompanyInvite[]> {
+    const sendToUser = await this.userRepository.findOne({
+      where: { id: userId },
+    });
+    if (!sendToUser)
+      throw new Error('no user found with that id for sending invite');
+    return this.companyInviteRepository.find({ where: { sendTo: sendToUser } });
   }
-  findInvitationsToUserByInvitationId(
+  async findInvitationsToUserByInvitationId(
     userId: number,
     invitationId: number,
   ): Promise<CompanyInvite | null> {
+    const sendToUser = await this.userRepository.findOne({
+      where: { id: userId },
+    });
+    if (!sendToUser)
+      throw new Error('no user found with that id for sending invite');
     return this.companyInviteRepository.findOne({
-      where: { id: invitationId, sendTo: userId },
+      where: { id: invitationId, sendTo: sendToUser },
     });
   }
   async acceptInvitation(
@@ -75,22 +88,41 @@ export class UsersService {
   ) {
     const companyInvite = await this.companyInviteRepository.findOne({
       where: { id: invitationId },
+      relations: ['sendTo', 'companyId'],
     });
+
     if (!companyInvite) return 'No such invite exists';
+
+    if (companyInvite.sendTo.id !== userId) {
+      return 'You are not authorized to respond to this invitation';
+    }
+
+    const user = await this.userRepository.findOne({ where: { id: userId } });
+    if (!user) return 'No user exists against that ID';
+
+    const company = await this.companyRepository.findOne({
+      where: { id: companyInvite.companyId.id },
+    });
+    if (!company) return 'No such company exists';
+
     if (acceptInvitation === 'accept') {
-      const { companyId, role } = companyInvite;
+      const { role } = companyInvite;
+
       const userCompany = new UserCompany();
-      userCompany.userId = userId;
-      userCompany.companyId = companyId;
+      userCompany.userId = user;
+      userCompany.companyId = company;
       userCompany.role = role;
+
       await this.userCompanyRepository.save(userCompany);
       await this.companyInviteRepository.delete(invitationId);
-      return 'successfully accepted invitation';
+
+      return 'Successfully accepted invitation';
     } else {
       await this.companyInviteRepository.delete(invitationId);
-      return 'successfully rejected invitation';
+      return 'Successfully rejected invitation';
     }
   }
+
   private async hashString(str: string): Promise<string> {
     const saltRounds = 10;
     return await bcrypt.hash(str, saltRounds);
